user_service/requirements.txt
amqp==5.2.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.0.10
django-filter==23.3
djangogrpcframework==0.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
kafka-python==2.0.2
kombu==5.3.4
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
XlsxWriter==3.1.9
